Renamed new installer directory
