+++ .gitmodules
@@ -1,6 +1,6 @@
 [submodule "software/client/tests/googletest"]
 	path = software/client/tests/googletest
 	url = https://github.com/google/googletest.git
-[submodule "software/new-installer/tests/googletest"]
-	path = software/new-installer/tests/googletest
-	url = https://github.com/google/googletest.git
+[submodule "software/installer/tests/googletest"]
+	path = software/installer/tests/googletest
+	url = git@github.com:google/googletest.git
