+++ software/installer/resources/resources.qrc
@@ -1,0 +1,5 @@
+<RCC>
+    <qresource>
+        <file>easyPlay-client</file>
+    </qresource>
+</RCC>
