--- software/new-installer/.clang-format
@@ -1,5 +1,0 @@
-{
-  BasedOnStyle: Chromium,
-  ColumnLimit: 0,
-  AlignConsecutiveMacros: true,
-}
