--- software/new-installer/.vscode/settings.json
@@ -1,83 +1,0 @@
-{
-  "files.associations": {
-    "__bit_reference": "cpp",
-    "__bits": "cpp",
-    "__config": "cpp",
-    "__debug": "cpp",
-    "__errc": "cpp",
-    "__hash_table": "cpp",
-    "__locale": "cpp",
-    "__mutex_base": "cpp",
-    "__node_handle": "cpp",
-    "__split_buffer": "cpp",
-    "__threading_support": "cpp",
-    "__tree": "cpp",
-    "__tuple": "cpp",
-    "__verbose_abort": "cpp",
-    "any": "cpp",
-    "array": "cpp",
-    "atomic": "cpp",
-    "bit": "cpp",
-    "bitset": "cpp",
-    "cctype": "cpp",
-    "clocale": "cpp",
-    "cmath": "cpp",
-    "complex": "cpp",
-    "condition_variable": "cpp",
-    "cstdarg": "cpp",
-    "cstddef": "cpp",
-    "cstdint": "cpp",
-    "cstdio": "cpp",
-    "cstdlib": "cpp",
-    "cstring": "cpp",
-    "ctime": "cpp",
-    "cwchar": "cpp",
-    "cwctype": "cpp",
-    "deque": "cpp",
-    "exception": "cpp",
-    "forward_list": "cpp",
-    "fstream": "cpp",
-    "initializer_list": "cpp",
-    "iomanip": "cpp",
-    "ios": "cpp",
-    "iosfwd": "cpp",
-    "iostream": "cpp",
-    "istream": "cpp",
-    "limits": "cpp",
-    "list": "cpp",
-    "locale": "cpp",
-    "map": "cpp",
-    "memory": "cpp",
-    "mutex": "cpp",
-    "new": "cpp",
-    "optional": "cpp",
-    "ostream": "cpp",
-    "ratio": "cpp",
-    "set": "cpp",
-    "sstream": "cpp",
-    "stack": "cpp",
-    "stdexcept": "cpp",
-    "streambuf": "cpp",
-    "string": "cpp",
-    "string_view": "cpp",
-    "system_error": "cpp",
-    "thread": "cpp",
-    "tuple": "cpp",
-    "type_traits": "cpp",
-    "typeinfo": "cpp",
-    "unordered_map": "cpp",
-    "unordered_set": "cpp",
-    "variant": "cpp",
-    "vector": "cpp",
-    "__nullptr": "cpp",
-    "__string": "cpp",
-    "chrono": "cpp",
-    "compare": "cpp",
-    "concepts": "cpp",
-    "algorithm": "cpp",
-    "qcoreapplication": "cpp",
-    "filesystem": "cpp",
-    "qfile": "cpp",
-    "qdebug": "cpp"
-  }
-}
