--- software/new-installer/CMakeLists.txt
@@ -1,66 +1,0 @@
-cmake_minimum_required(VERSION 3.26)
-
-set(CMAKE_AUTOUIC ON)
-set(CMAKE_AUTOMOC ON)
-set(CMAKE_AUTORCC ON)
-
-set(CMAKE_CXX_STANDARD 17)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-set(PROJ_LIB easyPlay-installer-lib)
-
-set(SRC_FILES
-    src/clientHandler.cpp
-    src/ui/setupWizard.cpp
-    src/ui/introPage.cpp
-    src/firefoxHandler.cpp
-    src/ui/installationPage.cpp
-    src/logger.cpp)
-
-set(INCLUDE_FILES
-    include/clientHandler.h
-    include/ui/setupWizard.h
-    include/ui/introPage.h
-    include/firefoxHandler.h
-    include/ui/installationPage.h
-    include/logger.h)
-
-set(RESOURCE_FILES resources/resources.qrc)
-
-if(CMAKE_BUILD_TYPE MATCHES Debug)
-    add_compile_definitions(DEBUG_BUILD)
-endif()
-
-project(easyPlay-installer LANGUAGES CXX)
-
-find_package(Qt6 REQUIRED COMPONENTS Core Concurrent Widgets)
-
-include_directories(include)
-
-add_library(${PROJ_LIB} STATIC ${SRC_FILES} ${INCLUDE_FILES}) 
-add_executable(${PROJECT_NAME} src/main.cpp ${RESOURCE_FILES})
-
-target_link_libraries(${PROJ_LIB} Qt6::Core Qt6::Concurrent Qt6::Widgets)
-target_link_libraries(${PROJECT_NAME} ${PROJ_LIB})
-
-install(TARGETS ${PROJECT_NAME} LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR} RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR})
-
-# TODO: rm
-# add_custom_command(TARGET ${PROJECT_NAME} POST_BUILD COMMAND ${CMAKE_COMMAND} -E copy "${CMAKE_INSTALL_BINDIR}/"${PROJECT_NAME} "${CMAKE_SOURCE_DIR}/dist/"${PROJECT_NAME})
-
-add_custom_command(TARGET ${PROJECT_NAME} POST_BUILD COMMAND bash "${CMAKE_SOURCE_DIR}/scripts/copy-installer.sh" ${PROJECT_NAME})
-
-# TESTING
-enable_testing()
-
-set(TEST_EXE easyplay-installer-tests)
-set(TEST_SRC_FILES
-    tests/clientHandlerTest.cpp
-    tests/firefoxHandlerTest.cpp)
-
-add_subdirectory(tests/googletest)
-
-add_executable(${TEST_EXE} ${TEST_SRC_FILES} ${RESOURCE_FILES})
-target_link_libraries(${TEST_EXE} gtest_main ${PROJ_LIB})
-
-add_test(NAME ${TEST_EXE} COMMAND ${TEST_EXE})
