--- software/new-installer/include/ui/installationPage.h
@@ -1,29 +1,0 @@
-#ifndef INSTALLATION_PAGE
-#define INSTALLATION_PAGE
-
-#include <QGridLayout>
-#include <QLabel>
-#include <QMessageBox>
-#include <QPushButton>
-#include <QWizardPage>
-
-class InstallationPage : public QWizardPage {
-  Q_OBJECT
- public:
-  explicit InstallationPage(QWidget* parent = nullptr);
-
- private:
-  QGridLayout* layout;
-  QLabel* label_dirPickerHeader;
-  QLabel* label_selectedDirPath;
-  QPushButton* btn_selectDestDir;
-  QString clientDestDir;
-
-  void initializePage() override;
-  bool validatePage() override;
-  void setUpGui();
-  void handleEvents();
-  void showPopup(const QString& text, QMessageBox::Icon icon, const QString& informativeText = "");
-};
-
-#endif  // INSTALLATION_PAGE
