--- software/new-installer/resources/resources.qrc
@@ -1,5 +1,0 @@
-<RCC>
-    <qresource>
-        <file>easyPlay-client</file>
-    </qresource>
-</RCC>
